+++ backend/src/main/java/com/eticaret/backend/repository/CategoryRepository.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/repository/CategoryRepository.java
 package com.eticaret.backend.repository;
 
 import com.eticaret.backend.model.Category;
@@ -7,8 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     Optional<Category> findByName(String name);
-
 }
