+++ backend/src/main/java/com/eticaret/backend/repository/GenderRepository.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/repository/GenderRepository.java
 package com.eticaret.backend.repository;
 
 import com.eticaret.backend.model.Gender;
@@ -7,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface GenderRepository extends JpaRepository<Gender, Integer> {
-
+public interface GenderRepository extends JpaRepository<Gender, Long> {
     Optional<Gender> findByName(String name);
 }
