+++ backend/src/main/java/com/eticaret/backend/repository/OrderRepository.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/repository/OrderRepository.java
 package com.eticaret.backend.repository;
 
 import com.eticaret.backend.model.Order;
@@ -7,12 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
-    /**
-     * Belirli bir kullanıcıya ait tüm siparişleri getirir.
-     * @param userId Kullanıcının ID'si
-     * @return Kullanıcının sipariş listesi
-     */
+public interface OrderRepository extends JpaRepository<Order, Long> {
     List<Order> findByUserUserId(Long userId);
 }
