+++ backend/src/main/java/com/eticaret/backend/repository/TagRepository.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/repository/TagRepository.java
 package com.eticaret.backend.repository;
 
 import com.eticaret.backend.model.Tag;
@@ -7,7 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Integer> {
-    // Optional ile dönmesini sağlıyoruz
+public interface TagRepository extends JpaRepository<Tag, Long> {
     Optional<Tag> findByName(String name);
 }
