+++ backend/src/main/java/com/eticaret/backend/repository/UserRepository.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/repository/UserRepository.java
 package com.eticaret.backend.repository;
 
 import com.eticaret.backend.model.User;
@@ -6,11 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-
-    // Kullanıcıyı username ile bulmak için
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
-
-    // Kullanıcıyı email ile bulmak için
     Optional<User> findByEmail(String email);
 }
