+++ backend/src/main/java/com/eticaret/backend/service/CategoryService.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/service/CategoryService.java
 package com.eticaret.backend.service;
 
 import com.eticaret.backend.model.Category;
@@ -14,33 +15,10 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
-    // Tüm kategorileri getir
-    public List<Category> getAllCategories() {
-        return categoryRepository.findAll();
-    }
-
-    // ID'ye göre kategori getir
-    public Optional<Category> getCategoryById(Integer id) {
-        return categoryRepository.findById(id);
-    }
-
-    // Yeni kategori ekle
-    public Category createCategory(Category category) {
-        return categoryRepository.save(category);
-    }
-
-    // Kategori güncelle
-    public Category updateCategory(Category category) {
-        return categoryRepository.save(category);
-    }
-
-    // Kategori sil
-    public void deleteCategory(Integer id) {
-        categoryRepository.deleteById(id);
-    }
-
-    // İsim ile arama (opsiyonel)
-    public Optional<Category> getCategoryByName(String name) {
-        return categoryRepository.findByName(name);
-    }
+    public List<Category> getAllCategories() { return categoryRepository.findAll(); }
+    public Optional<Category> getCategoryById(Long id) { return categoryRepository.findById(id); }
+    public Category createCategory(Category category) { return categoryRepository.save(category); }
+    public Category updateCategory(Category category) { return categoryRepository.save(category); }
+    public void deleteCategory(Long id) { categoryRepository.deleteById(id); }
+    public Optional<Category> getCategoryByName(String name) { return categoryRepository.findByName(name); }
 }
