+++ backend/src/main/java/com/eticaret/backend/service/TagService.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/service/TagService.java
 package com.eticaret.backend.service;
 
 import com.eticaret.backend.model.Tag;
@@ -14,33 +15,10 @@ public class TagService {
     @Autowired
     private TagRepository tagRepository;
 
-    // Tüm tag'ları getir
-    public List<Tag> getAllTags() {
-        return tagRepository.findAll();
-    }
-
-    // ID ile tag getir
-    public Optional<Tag> getTagById(Integer id) {
-        return tagRepository.findById(id);
-    }
-
-    // Yeni tag oluştur
-    public Tag createTag(Tag tag) {
-        return tagRepository.save(tag);
-    }
-
-    // Tag güncelle
-    public Tag updateTag(Tag tag) {
-        return tagRepository.save(tag);
-    }
-
-    // Tag sil
-    public void deleteTag(Integer id) {
-        tagRepository.deleteById(id);
-    }
-
-    // İsme göre tag getir
-    public Optional<Tag> getTagByName(String name) {
-        return tagRepository.findByName(name);
-    }
+    public List<Tag> getAllTags() { return tagRepository.findAll(); }
+    public Optional<Tag> getTagById(Long id) { return tagRepository.findById(id); }
+    public Tag createTag(Tag tag) { return tagRepository.save(tag); }
+    public Tag updateTag(Tag tag) { return tagRepository.save(tag); }
+    public void deleteTag(Long id) { tagRepository.deleteById(id); }
+    public Optional<Tag> getTagByName(String name) { return tagRepository.findByName(name); }
 }
