+++ backend/src/main/java/com/eticaret/backend/service/UserService.java
@@ -1,3 +1,4 @@
+// src/main/java/com/eticaret/backend/service/UserService.java
 package com.eticaret.backend.service;
 
 import com.eticaret.backend.model.User;
@@ -14,37 +15,11 @@ public class UserService {
     @Autowired
     private UserRepository userRepository;
 
-    // Tüm kullanıcıları getir
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    // ID'ye göre kullanıcı getir
-    public Optional<User> getUserById(Integer id) {
-        return userRepository.findById(id);
-    }
-
-    // Yeni kullanıcı ekle
-    public User createUser(User user) {
-        return userRepository.save(user);
-    }
-
-    // Kullanıcı güncelle
-    public User updateUser(User user) {
-        return userRepository.save(user);
-    }
-
-    // Kullanıcı sil
-    public void deleteUser(Integer id) {
-        userRepository.deleteById(id);
-    }
-
-    // E-posta veya kullanıcı adı ile arama (isteğe bağlı)
-    public Optional<User> getUserByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
-
-    public Optional<User> getUserByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
+    public List<User> getAllUsers() { return userRepository.findAll(); }
+    public Optional<User> getUserById(Long id) { return userRepository.findById(id); }
+    public User createUser(User user) { return userRepository.save(user); }
+    public User updateUser(User user) { return userRepository.save(user); }
+    public void deleteUser(Long id) { userRepository.deleteById(id); }
+    public Optional<User> getUserByEmail(String email) { return userRepository.findByEmail(email); }
+    public Optional<User> getUserByUsername(String username) { return userRepository.findByUsername(username); }
 }
